Fold copy&pasted parts of Button / AboutDialog / GpsPosition admin types onto new BaseButtonType
