+++ src/Mapbender/CoreBundle/Element/Type/AboutDialogAdminType.php
@@ -3,18 +3,13 @@ namespace Mapbender\CoreBundle\Element\Type;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class AboutDialogAdminType extends AbstractType
 {
-    /**
-     * @inheritdoc
-     */
-    public function configureOptions(OptionsResolver $resolver)
+
+    public function getParent()
     {
-        $resolver->setDefaults(array(
-            'application' => null
-        ));
+        return 'Mapbender\CoreBundle\Element\Type\BaseButtonAdminType';
     }
 
     /**
@@ -22,15 +17,10 @@ class AboutDialogAdminType extends AbstractType
      */
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
-        $builder
-            ->add('label', 'Symfony\Component\Form\Extension\Core\Type\CheckboxType', array(
-                'required' => false,
-                'label' => 'mb.core.admin.button.show_label',
-            ))
-            ->add('tooltip', 'Symfony\Component\Form\Extension\Core\Type\TextType', array(
-                'required' => false,
-            ))
-        ;
+        // Icon is hard-coded, remove upstream icon field.
+        // @todo: allow configuration, after providing the previously hard-coded setting as a default
+        if ($builder->has('icon')) {
+            $builder->remove('icon');
+        }
     }
-
 }
