+++ src/Mapbender/CoreBundle/Element/Type/BaseButtonAdminType.php
@@ -1,0 +1,40 @@
+<?php
+
+
+namespace Mapbender\CoreBundle\Element\Type;
+
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class BaseButtonAdminType extends AbstractType
+{
+    /**
+     * @inheritdoc
+     */
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver->setDefaults(array(
+            'application' => null,
+        ));
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function buildForm(FormBuilderInterface $builder, array $options)
+    {
+        $builder
+            ->add('tooltip', 'Symfony\Component\Form\Extension\Core\Type\TextType', array(
+                'required' => false,
+            ))
+            ->add('icon', 'Mapbender\CoreBundle\Element\Type\IconClassType', array(
+                'required' => false,
+            ))
+            ->add('label', 'Symfony\Component\Form\Extension\Core\Type\CheckboxType', array(
+                'required' => false,
+                'label' => 'mb.core.admin.button.show_label',
+            ))
+        ;
+    }
+}
