+++ src/Mapbender/CoreBundle/Element/Type/ControlButtonAdminType.php
@@ -1,14 +1,17 @@
 <?php
+
+
 namespace Mapbender\CoreBundle\Element\Type;
+
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
 
-class ButtonAdminType extends AbstractType
+class ControlButtonAdminType extends AbstractType
 {
     public function getParent()
     {
-        return 'Mapbender\CoreBundle\Element\Type\ControlButtonAdminType';
+        return 'Mapbender\CoreBundle\Element\Type\BaseButtonAdminType';
     }
 
     /**
@@ -17,16 +20,13 @@ class ButtonAdminType extends AbstractType
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
         $builder
-            ->add('click', 'Symfony\Component\Form\Extension\Core\Type\TextType', array(
+            ->add('group', 'Symfony\Component\Form\Extension\Core\Type\TextType', array(
                 'required' => false,
             ))
-            ->add('action', 'Symfony\Component\Form\Extension\Core\Type\TextType', array(
-                'required' => false,
-            ))
-            ->add('deactivate', 'Symfony\Component\Form\Extension\Core\Type\TextType', array(
+            ->add('target', 'Mapbender\CoreBundle\Element\Type\TargetElementType', array(
+                'application' => $options['application'],
                 'required' => false,
             ))
         ;
     }
-
 }
