+++ src/Mapbender/ManagerBundle/Resources/views/Element/gpsposition.html.twig
@@ -1,9 +1,6 @@
 <div class="elementForm wideLabelForm">
     {{ form_row(form.configuration.label) }}
     {{ form_row(form.configuration.autoStart) }}
-    <div class="clearContainer"></div>
     {{ form_row(form.title) }}
-    {{ form_row(form.configuration.tooltip) }}
-    {{ form_row(form.configuration.icon) }}
     {{ form_rest(form.configuration) }}
 </div>
